+++ gazebo/msgs/msgs.cc
@@ -953,8 +953,16 @@ namespace gazebo
       else if (geomElem->GetName() == "heightmap")
       {
         result.set_type(msgs::Geometry::HEIGHTMAP);
-        msgs::Set(result.mutable_heightmap()->mutable_size(),
-            geomElem->Get<ignition::math::Vector3d>("size"));
+
+        // We do not want to set the size field to be the default values of
+        // [1, 1, 1] if not specified (size is optional for DEMs). So mark it as
+        // zero for now.
+        // TODO remove the required rule in heightmapgeom.proto's size field
+        ignition::math::Vector3d size;
+        if (geomElem->HasElement("size"))
+          size =  geomElem->Get<ignition::math::Vector3d>("size");
+        msgs::Set(result.mutable_heightmap()->mutable_size(), size);
+
         msgs::Set(result.mutable_heightmap()->mutable_origin(),
             geomElem->Get<ignition::math::Vector3d>("pos"));