+++ Changelog.md
@@ -2,7 +2,10 @@
 
 ## Gazebo 7.x.x (2016-xx-xx)
 
-1. Update QT render loop to throttle based on UserCamera::RenderRate. 
+1. Fix DEM heightmap size, collision, scale
+    * [Pull request 2477](https://bitbucket.org/osrf/gazebo/pull-request/2477)
+
+1. Update QT render loop to throttle based on UserCamera::RenderRate.
     * [Pull request 2476](https://bitbucket.org/osrf/gazebo/pull-request/2476)
     * [Issue 1560](https://bitbucket.org/osrf/gazebo/issues/1560)